Moved tests for `NoEntityFrameworkTests` into a different database, because two variants of `TodoItem` exist (where one is a subset of the other), causing tests to fail/succeed depending on their execution order.
A separate database is needed instead of schema, because `context.Database.EnsureCreated()` only checks for database existence and does not create additional tables when the database already exists.

Furthermore, I decided to remove fields from `NoEntityFrameworkExample.Models.TodoItem` that do not exist in the INSERT SQL statement in `NoEntityFrameworkExample.Services.TodoItemService` and because non-nullable CreatedDate causes insert to fail.