Optimized/fixed TypeLocator.GetGenericInterfaceImplementation
This method was identified as a hot path in profiling JsonApiDotNetCoreExample tests.

Type.GetGenericTypeDefinition() on the incoming parameter was called inside the loop, which is unneeded. And it compared generic type parameters before looking at the type itself, potentially evaluating too much.

This change makes it run 30% faster. This change also improves input validation and fixes the bug where only the first type parameter was matched against found types.