Rewrite of reading the request body and converting it into models.
The existing validation logic inside controllers, JsonApiReader and RequestDeserializer/BaseDeserializer has been consolidated into a set of adapters that delegate to each other, passing along the current state. The dependency on `HttpContext` has been removed and similar errors are now grouped, in preparation to unify them. Because in this commit we always track the location of errors and write them to `error.source.pointer`, the error messages can be unified and simplified in future commits.

In this commit, I've gone through great lenghts to preserve the existing error messages in our tests as much as possible, while keeping all tests green. All error tests in ReadWrite and Operations now have assertions on the source pointer. Added tests for missing/invalid 'data' in resource requests. Removed outdated unit tests and added new one for handling attributes of various CLR types.

Updated benchmark code. The synthetic BenchmarkDotNet reports it has become 3-7% slower (we're talking microseconds here) compared to the old code. But when running full requests with our pipeline-instrumentation turned on, the difference falls in the range of background noise and is thus unmeasurable.

```
BEFORE:
          Read request body ...............................  0:00:00:00.0000167 ...   1.08%  } 167 + 434 = 601
            JsonSerializer.Deserialize ....................  0:00:00:00.0000560 ...   3.62%  +
            Deserializer.Build (single) ...................  0:00:00:00.0000434 ...   2.80%  }

AFTER:
          Read request body ...............................  0:00:00:00.0000511 ...   3.43%
            JsonSerializer.Deserialize ....................  0:00:00:00.0000537 ...   3.61%
```