Added model and tests that represent the scenario described at https://learn.microsoft.com/en-us/ef/core/saving/cascade-delete#database-cascade-limitations
The article states various limitations in SQL Server:

> Some databases, most notably SQL Server, have limitations on the cascade behaviors that form cycles.

> Using cascading deletes and cascading nulls in the database at the same time will almost always result in relationship cycles when using SQL Server.

When all related entities are loaded in the EF Core change tracker upfront, EF Core is able to handle the scenario by issuing multiple SQL statements.
And that's why the default delete behavior for optional relationships is `ClientSetNull` instead of `SetNull`.

The tests in this commit show that PostgreSQL handles the presented scenario successfully. Letting the database handle cascading is a lot more efficient.